support floats in the console
There are some strange bugs with parsing like !set type all 3.5 or !set temp all moo but they aren't too important ...

+++ src/lua/TPTSTypes.cpp
@@ -86,12 +86,12 @@ AnyType::~AnyType()
 
 //Number Type
 
-NumberType::NumberType(int number): AnyType(TypeNumber, ValueValue())
+NumberType::NumberType(float number): AnyType(TypeNumber, ValueValue())
 {
 	value.num = number;
 }
 
-int NumberType::Value()
+float NumberType::Value()
 {
 	return value.num;
 }
